(139) Moved tests into the package directories... Deleted workflow
