--- .github/workflows/go.yml
@@ -1,25 +1,0 @@
-name: Go
-
-on:
-  push:
-    branches: [ "main", "stable" ]
-  pull_request:
-    branches: [ "main", "stable" ]
-
-jobs:
-
-  build:
-    runs-on: ubuntu-latest
-    steps:
-    - uses: actions/checkout@v3
-
-    - name: Set up Go
-      uses: actions/setup-go@v3
-      with:
-        go-version: 1.18
-
-    - name: Build
-      run: go build -v ./...
-
-    - name: Test
-      run: go test -v ./...
