+++ data/data_test.go
@@ -1,18 +1,16 @@
-package tests
+package data
 
 import (
 	"testing"
-
-	"github.com/samjtro/go-tda/data"
 )
 
 func TestData(t *testing.T) {
-	_, err := data.RealTime("AAPL")
+	_, err := RealTime("AAPL")
 
 	if err != nil {
 		t.Fatalf(err.Error())
 	}
-	_, err = data.PriceHistory("AAPL", "month", "1", "daily", "1")
+	_, err = PriceHistory("AAPL", "month", "1", "daily", "1")
 
 	if err != nil {
 		t.Fatalf(err.Error())
