+++ instrument/instrument_test.go
@@ -1,19 +1,17 @@
-package tests
+package instrument
 
 import (
 	"testing"
-
-	"github.com/samjtro/go-tda/instrument"
 )
 
 func TestInstrument(t *testing.T) {
-	_, err := instrument.Simple("AAPL")
+	_, err := Simple("AAPL")
 
 	if err != nil {
 		t.Fatalf(err.Error())
 	}
 
-	_, err = instrument.Fundamental("AAPL")
+	_, err = Fundamental("AAPL")
 
 	if err != nil {
 		t.Fatalf(err.Error())
