+++ movers/movers_test.go
@@ -1,13 +1,11 @@
-package tests
+package movers
 
 import (
 	"testing"
-
-	"github.com/samjtro/go-tda/movers"
 )
 
 func TestMovers(t *testing.T) {
-	_, err := movers.Get("$DJI", "up", "percent")
+	_, err := Get("$DJI", "up", "percent")
 
 	if err != nil {
 		t.Fatalf(err.Error())
